scraper/requirements.txt
Scrapy==2.11.0
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyyaml==6.0.1
python-levenshtein==0.23.0

